pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
